// src/apollo/alerts/nrt/loader/loader.types.ts
/*
 * Copyright 2021 OpenBA-NextGenTV Contributors (https://OpenBA-NextGenTV.tech)
 * Copyright 2021 Sinclair Broadcast Group, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type AlertType = {
  text: string;
  priority: number;
  expire: number;
  originalPublishTime: number;
  latestPublishTime: number;
  targets: string[];
  eventCode: string;

  linkAppInfo: {
    guid: string;
  };
};

export type AlertFeed = {
  guid: string;
  title: string;
  pages: FeedPage[];
};

export type FeedPage = {
  title: string;
  story: string;
  media: FeedMedia[];
};

export type FeedMedia = {
  caption: string;
  title: string;
  media: string;
  thumbnail: string;
};